fix(#148): 스크랩 북 헤더 고정
